Fixed non-ascii chars and latex deprecations
make errors resulting from non-ascii characters in python files in ./Scripts
were corrected by adding encoding information to the heads of the files.

make errors arising from non-ascii characters in latex and xml files were
corrected by using correct ascii quote marks and the latex command \pounds.

make errors arising from deprecated latex commands `\rm` and `\bf` were
corrected by replacement with `\textrm`, `\mathrm` or `\bfseries` as appropriate

+++ Scripts/basicgraph.py
@@ -88,10 +88,10 @@ class basicgraph:
                  graphstr += "edge[style=invis];\n"
                  graphstr += backnodes[len(backnodes)-2] + " -> " + self.nodes[j] + '\n'
              backnodes.append(self.nodes[j])
-      for i in range(0,math.floor(len(forwardnodes)/2)+1) :
+      for i in range(0,int(math.floor(len(forwardnodes)/2)+1)) :
           if len(forwardnodes)>2*i+2 :  
              graphstr += "{ rank=same; " + forwardnodes[2*i] + ";" + forwardnodes[2*i+1] + "; } \n" 
-      for i in range(0,math.floor(len(backnodes)/2)+1) :
+      for i in range(0,int(math.floor(len(backnodes)/2)+1)) :
           if len(backnodes)>2*i+2 : 
              graphstr += "{ rank=same; " + backnodes[2*i] + ";" + backnodes[2*i+1] + "; } \n"
       graphstr += "}"
